RAG Chatbot/requirements.txt
streamlit
faiss-cpu
numpy
langchain
langchain-huggingface
requests
langchain-community